+++ Game/release-notes.md
@@ -4,6 +4,7 @@
 
 - First release
 - Adding wrap on settings menu & difficulty menu
+- Adding icons to the menus
 
 ## Bug fixes